separate components from pages
