+++ src/components/admin/ShowModerator.tsx
@@ -5,16 +5,19 @@ import { ShowGroupRun } from '@/types/main';
 import BackArrow from '../shared/BackArrow';
 import { Song } from '@prisma/client';
 import ShowEditor from './ShowEditor';
+import { useRouter } from 'next/router';
 
 interface ShowModeratorProps {
   shows: ShowGroupRun[];
   todayShow: ShowWithVenueAndRun | undefined;
   allSongs: Song[];
-  back: () => void;
 }
 
-const ShowModerator: React.FC<ShowModeratorProps> = ({ shows, todayShow, allSongs, back }) => {
+const ShowModerator: React.FC<ShowModeratorProps> = ({ shows, todayShow, allSongs }) => {
+  const router = useRouter();
   const [selectedShow, setSelectedShow] = useState<ShowWithVenueAndRun | null>(null);
+
+  const back = () => router.back();
 
   return (
     <div className="flex flex-col items-center w-full max-w-md px-4">
