+++ src/components/data/songs/SongLookupContainer.tsx
@@ -10,16 +10,9 @@ import Head from 'next/head';
 import React, { useState } from 'react';
 import toast from 'react-hot-toast';
 
-interface SongLookupPageProps {
+export interface SongLookupContainerProps {
   allSongs: Song[];
 }
-
-export const getServerSideProps: GetServerSideProps<SongLookupPageProps> = async () => {
-  const allSongs = await getAllSongs();
-  return {
-    props: { allSongs },
-  };
-};
 
 /**
  * 
@@ -50,7 +43,7 @@ const SongInfoList = ({ song, data }: { song: Song; data: PhishNetSong }) => (
   </>
 );
 
-const SongLookupPage: React.FC<SongLookupPageProps> = ({ allSongs }) => {
+const SongLookupContainer: React.FC<SongLookupContainerProps> = ({ allSongs }) => {
   const [loading, setLoading] = useState(false);
   const [selectedSong, setSelectedSong] = useState<Song | null>(null);
   const [songData, setSongData] = useState<PhishNetSong | null>(null);
@@ -68,7 +61,6 @@ const SongLookupPage: React.FC<SongLookupPageProps> = ({ allSongs }) => {
       setLoading(false);
       return;
     }
-    console.log(songData);
     setSongData(songData);
     setSelectedSong(song);
     setLoading(false);
@@ -76,9 +68,6 @@ const SongLookupPage: React.FC<SongLookupPageProps> = ({ allSongs }) => {
 
   return (
     <div className="flex flex-col items-center">
-      <Head>
-        <title>Song Lookup | Go Phish</title>
-      </Head>
       <p className="text-title-regular my-5">Song Lookup</p>
       <div className="w-3/4 mb-5">
         <SongInput allSongs={allSongs} selectSong={selectSong} selectedSong={null} />
@@ -94,4 +83,4 @@ const SongLookupPage: React.FC<SongLookupPageProps> = ({ allSongs }) => {
   );
 };
 
-export default SongLookupPage;
+export default SongLookupContainer;
