+++ src/pages/guesses/run/[runId].tsx
@@ -1,34 +1,22 @@
-import GuessList from '@/components/guesses/GuessList';
-import ErrorMessage from '@/components/shared/ErrorMessage';
-import { RunInfo } from '@/components/shared/RunInfo';
-import { ResponseStatus } from '@/types/main';
-import BackArrow from '@/components/shared/BackArrow';
-import { getGuessesForRun } from '@/services/guess.service';
-import { getRunById, getRunWithVenue } from '@/services/run.service';
-import { getShowsForRun, getShowsForRunWithVenue } from '@/services/show.service';
-import { getUsersByIds } from '@/services/user.service';
-import { OrganizedGuesses, organizedGuessesForNight, organizeGuessesByUser } from '@/utils/guess.util';
-import { organizeArrayByField, parseObj } from '@/utils/utils';
-import { Guess, Run, Show, User } from '@prisma/client';
-import { GetServerSideProps } from 'next';
-import { useSession } from 'next-auth/react';
-import Head from 'next/head';
-import Link from 'next/link';
-import { useRouter } from 'next/router';
 import React from 'react';
-import { RunWithVenue } from '@/models/run.model';
-import { ShowWithVenue } from '@/models/show.model';
-import RadioGroup, { RadioOption } from '@/components/shared/RadioGroup';
-import { useThemeContext } from '@/store/theme.store';
+import Head from 'next/head';
+import { GetServerSideProps } from 'next';
+import ErrorMessage from '@/components/shared/ErrorMessage';
+import { ResponseStatus } from '@/types/main';
+import { getGuessesForRun } from '@/services/guess.service';
+import { getRunWithVenue } from '@/services/run.service';
+import { getShowsForRunWithVenue } from '@/services/show.service';
+import { getUsersByIds } from '@/services/user.service';
+import { organizeGuessesByUser } from '@/utils/guess.util';
+import { organizeArrayByField, parseObj } from '@/utils/utils';
+import { Guess } from '@prisma/client';
+import GuessRunContainer, { GuessRunContainerProps } from '@/components/guesses/run/GuessRunContainer';
 
-interface RunGuessContainerProps {
-  run?: RunWithVenue;
-  shows?: ShowWithVenue[];
-  guesses?: OrganizedGuesses;
+type GuessRunPageProps = Partial<GuessRunContainerProps> & {
   error?: string;
-}
+};
 
-export const getServerSideProps: GetServerSideProps<RunGuessContainerProps> = async (context) => {
+export const getServerSideProps: GetServerSideProps<GuessRunPageProps> = async (context) => {
   const runId = parseInt(context.params?.runId?.toString() || '');
   if (isNaN(runId)) {
     return {
@@ -36,7 +24,6 @@ export const getServerSideProps: GetServerSideProps<RunGuessContainerProps> = as
     };
   }
   let run = await getRunWithVenue(runId);
-  console.log(run);
   if (run === ResponseStatus.NotFound) {
     return {
       props: { error: 'Run not found.' },
@@ -72,82 +59,19 @@ export const getServerSideProps: GetServerSideProps<RunGuessContainerProps> = as
   };
 };
 
-const RunGuesses: React.FC<RunGuessContainerProps> = ({ run, shows, guesses, error }) => {
-  const { status } = useSession();
-  const router = useRouter();
-  const { night: runNight } = router.query;
-  const selectedNight = parseInt(runNight?.toString() || '');
-  const { color } = useThemeContext();
-
+const GuessRunPage: React.FC<GuessRunPageProps> = ({ run, shows, guesses, error }) => {
   if (error || !run || !shows || !guesses) {
     return <ErrorMessage error={error} />;
   }
-
-  if (status === 'loading') {
-    return <p>Loading...</p>;
-  }
-
-  const nightNumbers: number[] = shows.map((show) => show.runNight).sort();
-  const nightShow: Show | undefined = shows.find((s) => s.runNight === selectedNight);
-  const organizedGuesses = isNaN(selectedNight) ? guesses : organizedGuessesForNight(guesses, shows, selectedNight);
-
-  const chooseNight = (night: number | string) => {
-    if (night === 'total') {
-      router.query.night = '';
-      router.push(router, undefined, { shallow: true });
-    } else if (!isNaN(parseInt(night.toString()))) {
-      router.query.night = night.toString();
-      router.push(router, undefined, { shallow: true });
-    }
-  };
-
-  const nightRadioOptions: RadioOption[] = [...nightNumbers, null].map((nightNumber) => ({
-    value: nightNumber || 'total',
-    label: nightNumber === null ? 'Total' : `Night ${nightNumber}`,
-  }));
 
   return (
     <>
       <Head>
         <title>{run.name} - Guesses | Go Phish</title>
       </Head>
-      <div className="flex flex-col items-center pb-10">
-        <div className="flex justify-between items-center w-full max-w-500 p-4">
-          <div className="flex items-center space-x-2">
-            <BackArrow
-              width={16}
-              height={16}
-              link="/shows"
-              className="cursor-pointer flex items-center space-x-2"
-              svgClass="fill-black dark:fill-white"
-            >
-              <p className="">Shows</p>
-            </BackArrow>
-          </div>
-          <p className="text-2xl font-light">Guesses</p>
-          <div className="">
-            <Link
-              href={`/scores/run/${run.id}${nightShow ? `?night=${nightShow.runNight}` : ''}`}
-              className={`text-${color} my-2.5`}
-            >
-              Leaderboard
-            </Link>
-          </div>
-        </div>
-
-        <RunInfo run={run} large showLocation />
-        <RadioGroup
-          options={nightRadioOptions}
-          selected={selectedNight}
-          select={(nightNumber) => chooseNight(nightNumber)}
-          containerClass="mt-4"
-        />
-        <div className="w-full my-1">
-          <GuessList organizedGuesses={organizedGuesses} nightShow={nightShow || null} />
-        </div>
-      </div>
+      <GuessRunContainer run={run} shows={shows} guesses={guesses} />
     </>
   );
 };
 
-export default RunGuesses;
+export default GuessRunPage;
