+++ src/pages/scores/run/[runId].tsx
@@ -1,34 +1,22 @@
-import LeaderboardInfo from '@/components/scores/Leaderboard';
 import ErrorMessage from '@/components/shared/ErrorMessage';
-import { RunInfo } from '@/components/shared/RunInfo';
 import { ResponseStatus } from '@/types/main';
-import BackArrow from '@/components/shared/BackArrow';
 import { getGuessesForRun } from '@/services/guess.service';
-import { getRunById, getRunWithVenue } from '@/services/run.service';
-import { getShowsForRun, getShowsForRunWithVenue } from '@/services/show.service';
+import { getRunWithVenue } from '@/services/run.service';
+import { getShowsForRunWithVenue } from '@/services/show.service';
 import { getUsersByIds } from '@/services/user.service';
-import { RankedUserScores, rankedUserScoresForNight, rankScoresByUser } from '@/utils/guess.util';
+import { rankScoresByUser } from '@/utils/guess.util';
 import { organizeArrayByField, parseObj } from '@/utils/utils';
-import { Guess, Run, Show } from '@prisma/client';
+import { Guess } from '@prisma/client';
 import { GetServerSideProps } from 'next';
-import { useSession } from 'next-auth/react';
 import Head from 'next/head';
-import Link from 'next/link';
-import { useRouter } from 'next/router';
 import React from 'react';
-import { RunWithVenue } from '@/models/run.model';
-import { ShowWithVenue } from '@/models/show.model';
-import RadioGroup, { RadioOption } from '@/components/shared/RadioGroup';
-import { useThemeContext } from '@/store/theme.store';
+import RunScoreContainer, { RunScoreContainerProps } from '@/components/scores/run/RunScoreContainer';
 
-interface RunScoreContainerProps {
-  run?: RunWithVenue;
-  shows?: ShowWithVenue[];
-  rankedUserScores?: RankedUserScores;
+type RunScorePageProps = Partial<RunScoreContainerProps> & {
   error?: string;
-}
+};
 
-export const getServerSideProps: GetServerSideProps<RunScoreContainerProps> = async (context) => {
+export const getServerSideProps: GetServerSideProps<RunScorePageProps> = async (context) => {
   const runId = parseInt(context.params?.runId?.toString() || '');
   if (isNaN(runId)) {
     return {
@@ -71,83 +59,19 @@ export const getServerSideProps: GetServerSideProps<RunScoreContainerProps> = as
   };
 };
 
-const RunLeaderboard: React.FC<RunScoreContainerProps> = ({ run, shows, rankedUserScores, error }) => {
-  const { status } = useSession();
-  const router = useRouter();
-  const { night: runNight } = router.query;
-  const selectedNight = parseInt(runNight?.toString() || '');
-  const { color } = useThemeContext();
-
+const RunScorePage: React.FC<RunScorePageProps> = ({ run, shows, rankedUserScores, error }) => {
   if (error || !run || !shows || !rankedUserScores) {
     return <ErrorMessage error={error} />;
   }
-
-  if (status === 'loading') {
-    return <p>Loading...</p>;
-  }
-
-  const nightNumbers: number[] = shows.map((show) => show.runNight).sort();
-  const nightShow: Show | undefined = shows.find((s) => s.runNight === selectedNight);
-  const organizedScores = isNaN(selectedNight)
-    ? rankedUserScores
-    : rankedUserScoresForNight(rankedUserScores, shows, selectedNight);
-
-  const chooseNight = (night: number | string) => {
-    if (night === 'total') {
-      router.query.night = '';
-      router.push(router, undefined, { shallow: true });
-    } else if (!isNaN(parseInt(night.toString()))) {
-      router.query.night = night.toString();
-      router.push(router, undefined, { shallow: true });
-    }
-  };
-
-  const nightRadioOptions: RadioOption[] = [...nightNumbers, null].map((nightNumber) => ({
-    value: nightNumber === null ? 'total' : nightNumber,
-    label: nightNumber === null ? 'Total' : `Night ${nightNumber}`,
-  }));
 
   return (
     <>
       <Head>
         <title>{run.name} - Leaderboard | Go Phish</title>
       </Head>
-      <div id="leaderboard-page">
-        <div className="flex flex-col items-center pb-4">
-          <div className="flex justify-between items-center w-full max-w-500 p-4">
-            <div className="flex items-center space-x-2 ">
-              <BackArrow
-                width={16}
-                height={16}
-                link="/shows"
-                className="cursor-pointer flex items-center space-x-2"
-                svgClass="fill-black dark:fill-white"
-              >
-                <p className="">Shows</p>
-              </BackArrow>
-            </div>
-            <p className="text-2xl font-light">Leaderboard</p>
-            <div className="">
-              <Link
-                href={`/guesses/run/${run.id}${nightShow ? `?night=${nightShow.runNight}` : ''}`}
-                className={`text-${color} my-2.5`}
-              >
-                All Guesses
-              </Link>
-            </div>
-          </div>
-          <RunInfo run={run} large showLocation />
-          <RadioGroup
-            options={nightRadioOptions}
-            selected={selectedNight}
-            select={(nightNumber) => chooseNight(nightNumber)}
-            containerClass="mt-4"
-          />
-          <LeaderboardInfo rankedUserScores={organizedScores} />
-        </div>
-      </div>
+      <RunScoreContainer run={run} shows={shows} rankedUserScores={rankedUserScores} />
     </>
   );
 };
 
-export default RunLeaderboard;
+export default RunScorePage;
